bugfix, multiQ stopped when few prods/ many cons
When many consumers w/ few producers present, at breakdown time some
consumers were stuck in a dequeue loop on an empty queue with no hope of
escape. Added an iterruption point in the loop. Common case says that
the point will be rarely called, but it might slow down performance.